added test cases, found a problem in 30

+++ no_loop_problems/30/specification-fs.h
@@ -14,4 +14,4 @@
     assigns *out;
     ensures IsWithinMargin(l, p, q, *out);
 */
-void calculateSecondSpellCollisionDistance(int l, int p, int q, int *out);
+void calculateSecondSpellCollisionDistance(double l, double p, double q, double *out);
